blog/requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
attr==0.3.1
attrs==19.3.0
backports.zoneinfo==0.2.1
Django==3.2.5
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.0.1
promise==2.3
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
